Add property to cooperate with syntax-local-value
This commit introduces a new struct type property that
allows a struct to cooperate with syntax-local-value to redirect
the access to another identifier. This functionality is similar
to the cooperation that rename transformers enable.

The advantage of this property is that it can be paired with a
set!-transformer to enable expansion behavior that is different
from rename transformers.

In addition, since the syntax-local-value access can be redirected
via a procedure, the choice of identifier to redirect to can be
delayed until expansion time.